pyproject.toml
[tool.poetry]
name = "flask-ticket-system"
version = "0.1.0"
description = ""
authors = ["Dmytro Maior <mayor4090Gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kytool = "^0.5.2"
flask = "^3.0.3"
pyjwt = "^2.8.0"
pydantic = "^2.8.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.11.0"
pre-commit = "^3.7.1"
pytest = "^8.3.1"
ipykernel = "^6.29.5"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
